Refactored query into an extension method on DbContext.
